requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
requests
pypdf
fpdf